Temporarily redirect all unauthenticated users
I intend to use this to hide what I'm working on for the time being. It
probably should be based on environment for now; my idea is to have a
coming-soon page up in front rather than my domain's registrar page.

Will see if there's a nicer way to do this once I have the page
designed.

May also make this a layout useful for single message displays, such as
a maintenance page. That could be handy.