Reformat error page layout
Passing the status body to as data was unnecessary. Better to have an actual body section of the status page that can be customized per status page.

Also decided the 'extras' section was better named `status-footer`, and it was also better to wrap the home link in a paragraph element instead of having a random link hanging around in the layout.