Add padding around status page content
This prevents the content from riding the bottom of the page as it was a moment ago.

This was missed because I previously forced the footer to sit at the bottom of the window. I may still want to make this happen somehow, but I'll have to come back to it.