Setup cards to be clickable
This is actually a Tweet-like feature. Rather than wrapping each card in an anchor element, each card will set its own click event handler and decide whether to make the cursor change to a pointer on mouse over.

But if the `url` prop isn't passed, then the cursor will be unaffected and the click handler will simply do nothing.