Prepare to display single-tweet "timeline"
This refers to reusing the Twitter timeline to display the single newest tweet.

Since it won't actually be a timeline, none of the graphical stuff related to the timeline will be displayed. Only the most recent tweet will be displayed.

This is only the prep work for this though -- both displaying only 1 tweet, and hiding the timeline's graphical elements.