Setup dummy GitHub Activity Item markup
This is for visualizing what will be displayed according to what data is given to the component (since there will be multiple types of GitHub events).

This first set of dummy markup is for push events. Others will be built out once the basics of these events on the timeline are built out.