Remove unnecessary global references for Twitter timeline
The HTML ID would be useless since this component is meant to be used in more than 1 place (even if it may only be no more than 1 time on each page it appears).

Also, the above applies to setting a global reference to the current TwitterTimeline instance (not to mention it was only for testing anyway).