Reformat template for better event message rendering
This allows the message selection to be more automatic. It's not 100% ideal, but it's better than creating nearly duplicate sets of markup for the various messages from Create, Push and Delete events.

Doing this did require me to extract the sentence parts into objects containing lists of prepositions and actions used in each event summary, with duplicate field names. Will look later to see if there's a better way to do this.

This -- and a conditional technique in Vue using `<template>` elements -- also allowed me to remove the wrapper div from around body of the rendered event.

Additionally, add new `branchName` and `branchUrl` properties for building the name of the branch and its related URL for the GitHub repo, and remove the `mounted` method since it shouldn't be needed in the future.