Style and link various parts of component
Changes:
- Reformat markup to make it more readable
- Add computed properties and a new method for generating various URLs
- Fix how branch URLs are generated
- Invert event-related data organization
- Style timeline marker icons

The second-to-last item refers to taking the icons, actions, and prepositions properties and inverting how they're structured. Now there will be a property of strings for each type of event instead.

I'll have to see about a better way to handle that, but this makes it easier to allow multiple types of events to share some strings -- particularly, "action" and "preposition" words used in each event summary.

Also, the branch URL computed property needed to be fixed; clicking the links created using these URLs would bring the user to the related GitHub API endpoint rather than the actual page on GitHub for the branch. This now correctly brings the user to the tree view for the branch on GitHub.