Add prop for deciding position of timeline line
This will allow adjusting the position of the line for different purposes -- especially in the case of this GitHub activity feed. Since the activity icons are much smaller than, say, the Twitter icons, it makes sense to not force them to take up the same width.

Making this change also required converting the `commonClasses` data property to a computed property. This allows the computed property to read the prop value into the returned string of classes.