Bulk-add render support for 2 new event types
The new types supported are IssuesEvent and IssueCommentEvent.

Now that these 2 are supported, I definitely am starting to think it worthwhile to split all of this functionality and state management into separate components. I'll do that after handling an email notification system for new event types that support hasn't been added for yet.