Begin sending alerts when an unsupported event type is found
This SHOULD work, though I haven't had a chance to test it yet.

I basically had to add a "to" address and name to the mail config that would read from related environment vars. I then create an empty array in the model and store the address information in the `alertRecipients` array.

To do the checking properly, I decided to store the current event's type earlier on in the loop. This allows the filtering cycle to be skipped for that item if it's an unsupported event type.

Then, once the loop completes, the email alert is sent with a linked list of event types and the activity data is returned to the caller.