Fix issues causing 500 error when attempting to load GitHub activity
First issue: The import of the `GitHubEventEmail` class wasn't properly qualified. Namespace references are assumed to be done from the root level, not from the level of the current namespace (`App`, in this case).

Second issue: the `Mail` class being used to send the email exists within the global namespace. Since we're working within the `App` namespace currently, a backslash is needed for this one case. Alternatively, I believe another `use` line near the top would be enough to import it into this namespace, but this is all that's needed since `Mail` is only used once.

Third issue: the model was attempting to send an email with an empty array, which means there were no new event types to worry about. The model will no longer attempt to send this type of email unless there's at least 1 new event type in the array.