Stop displaying more than 4 commits for push events
Since these can get very long, especially for merge commits, I've decided to cut the number displayed down to the first 4 in the list. If there are any others, they will be counted and a link will be displayed that will take the visitor to the list of commits for the particular branch.