Modify logo row layout somewhat
Previously, the logos used a 1/2 layout on medium range screen sizes and a 1/6-1/3-1/3-1/6 layout on large size screens and up. This was due to only having 2 logos in the row before.

Since there are now 3 logos, making the row have a 1/3-1/3-1/3 layout on medium screens and up makes more sense.