Create separate Vue components for each event type
Each has been cleaned of any extra conditional, and otherwise useless, code that would have been used in case of other event types.

This is only a starting point. Next, the most common factors need to be identified so that there won't be *too much* repeating of code... The goal is to prevent repeating code if possible, but avoid having a single component that keeps getting more an more bloated.

Will see where this leads.