Update/add some dependencies
Dependencies updated/added:
- laravel-mix 2.1.14 -> 3.0.0
- har-validator ~5.1.3

The first item happens to also include a fix for a known vulnerability.

The second dependency "added" is actually a manual update of a dependency of another dependency, ad nauseam. NPM was looking for version 5.1.2, but it wasn't available. So I manually installed 5.1.3 instead, which seemed to satisfy the eventual needs of `node-sass` (down the dep. tree).