Support repo create events in CreateEvent component
If the create event was for a repo instead of a ref within the repo, the component would end up not displaying. This is because the `payload.ref` data is `null` in this case, causing an error in one of the computed properties.

Wrapping the relevant markup inside a `v-if` block prevents the unnecessary computed property from being computed in the first place and also hides content unrelated to a repo being created.