Handle PullRequestEvent data transformations inline
Just learned about the `transform` method for use with Eloquent collections. Using this method in tandem with the `filter` method allows stripping out unwanted data and then transforming the remaining data as needed.

In this case, it allows me to have nested filters inline, reducing the number of variables used to reference various collections to 0.