Roll back some dependencies and install node-sass
The reason for this is that Laravel Mix 4.x (or possibly Webpack 4) is causing a few issues with generated public assets.

In this case, some files that are not chunked modules end up within the same path as the chunked modules, and the generated CSS file ends up filled with JS instead of CSS. Rolling back to what I had before will fix the issue for now.

Changes:
- Revert `laravel-mix` to `3.0.0`
- Revert `laravel-mix-purgecss` to `3.0.0`
- Remove `resolve-url-loader`
- Remove `sass`
- Install `node-sass` version `4.11.0`

Also, I was previously using a globally-installed version of `node-sass`, but it's probably best to just include the NPM module as a direct dev dependency since it's required for the above version of `laravel-mix`.