Remove some unnecessary single quotes and add white color back in
Changes:
- Remove single quotes from around object keys that don't need them
- Add the "white" color back in

The latter fixes an issue with Tailwind not being able to find the color white. I originally misunderstood some instructions when I was removing stuff that was "unchanged". It turns out that those instructions were referring to removing entire config sections that were unchanged.

In other words, if I had changes to the colors object *at all*, then that colors object would override the one pulled in from the default config. So `white` needs to be in there.