Fix some text sizing issues
Most of this involves explicitly setting a font size on headings. Tailwind zeroes these font sizes out by default to intentionally force devs to set their own instead of relying on the browser to set one for us.

There is also a change in the header to fix the issue with the size of the logo text.