Use default Tailwind letterSpacing and add TODO comment
I still had my old letterSpacing scale from the development version of Tailwind. The default scale hasn't changed drastically, but it does allow greater flexibility. I've updated an instance of one of the `tracking` classes also to make use of this change.

I've also added a TODO to look into the new default screen size breakpoints. As stated in the upgrade guide, none of the screen sizes has decreased, so there shouldn't be issues. But if I decide to go with the default Tailwind scale, they'll need to be tested. The previous defaults did work fine previously though.