Fix bug preventing middlewares to redirect when disabled
Both the Coming Soon and Maintenance Mode middlewares are supposed to redirect to the homepage under certain conditions. One of those conditions is that the environment is not set to "local" (as in, for local development). This should be happening also when both middlewares have their related environment vars set to `false`.

This was happening because I was passing their related route *paths* to the `Request::is()` method instead of the route *names*. Passing the route *names* to this method fixes the issue preventing a redirect home when these my "Coming Soon" and "Maintenance Mode" are supposed to be disabled in a non-local environment.