Switch to using timeline partial as a component
Before part of my HTML was closing automatically before the section from the timeline partial would render. I was thinking I could render the timeline partial into this HTML forgetting that the twitter-timeline partial was just extending the timeline partial--meaning, using it as a layout. This isn't the same thing as working with components, like in Vue.

Luckily, Laravel has components with slots. Using this feature allows me to import a partial into another partial and render content into that partial directly.