Add a small hack to fix a small difference in alignment
This difference occurs because of how rendering is handled in Vue. Basically, Vue is outright replacing the timeline sections with their respective timelines, not inserting them into the root elements.

So it's really Vue that's causing the alignment difference, but this small hack will at least make them uniform for now.