Fix commitSubject method display issues
Changes:
- Fix search for first line break in `msg`
- Add condition to ensure `$msgEnd` isn't considered "empty"
- Format ternary expression for easier readability

For the first issue, `\n` needs to be double-quoted instead of single-quoted to be taken as a literal line-break, otherwise it's read as the literal characters "\n".

For the second issue, it wasn't enough to check that `$msgEnd` was greater than `0` since there was no guarantee that it would be a numeric value.