Combine 2 `class` attributes on same HTML element
This was a mistake caused by how the original markup was set up in the Vue template. In Vue, if there's a hard-coded attribute and an attribute binding with the same name, Vue will combine them after evaluating the bound attribute. This obviously won't work in Blade, which was causing the icon to not display at all.