Fix some potential issues in a few activity partials
Issues include how actions and other settings are set, as well as what some of the icon/action/preposition variables are set to (some still had incorrect values from being copied from a different source).

For the Issues Event, the icon and action need to be conditionally set.

For the Pull Request Event, the action needs to be conditionally set. This also simplifies the reference used in the switch statement.