Generate source maps for debugging in production
I shouldn't see a performance hit unless I have my dev tools open anyway, so this shouldn't be a problem.

The Laravel Mix docs mention that there will be a performance hit, but they're very vague about when that actually happens. Reading up on it further, it seems that it's only when the dev tools are open; source maps are not downloaded when doing regular browsing.