Play around with varying indentations in .editorconfig
Well...this explains why my Vue files were constantly changing to 4 spaces for indentation. I forgot I had a .editorconfig that was set to 4 for everything except for Markdown and YAML.

Will have to play with this more later to see what makes sense. Markup files for sure need to be 2 spaces though.