Couple of minor projects-list template fixes
1. Wrap assignment options in parentheses for casting
2. Remove unnecessary semi-colon from blade syntax

The second fix is necessary to prevent accidentally attempting to cast a
'null' value to an 'int'. This results in an issue that using the `??`
operator was meant to fix.

Regarding the second fix, these semi-colons are only needed for actual
PHP anyway, so putting them at the ends of Blade lines causes them only
to be rendered.