Remove custom Maintenance Mode
This removes the related middleware, route, HTTP Kernel entries, and app config entry.

My reason for doing this is that Laravel has a built-in maintenance mode I can take advantage of. I plan on doing that, and possibly reusing my maintenance mode view as my 503 error template (the 503 error being the one used for maintenance mode).