Fix exception message conditions
I may need to find a better place to do this. There are cases in which
this exception object is not set (such as with manually looking at these
error views), and cases in which there is no message in the exception
object.

For now I've added checks to each error view individually.