Hide GA popup based on custom cookie
Initially decided one of the automatic popup hiding conditions should be
based on the `DNT` cookie being set to `0`.  Forgot that this cookie
will initially be set this way automatically before the user either
allows or denies tracking.

A better solution seems to be having the popup set its own cookie
indicating that the user has interacted with it already.