Few small changes to error route
Changes:
- Use `where` clause in route definition
- Remove manual check for null or empty `code` parameter value
- Use `abort` to properly load error views with exceptions

The first 2 changes are related. With this `where` clause, the route
will only run if the `code` paramater has a value that matches the given
expression, so there's no longer a need for manual checks to ensure this
paramater is set properly.

The third change forces Laravel to actually throw the appropriate
exception for a given error code. Even though it's being faked, this
ensures the proper error data is passed to the related view.