Update some head/footer section names/references
The gist of this is that when I was having individual page views use
something like `@section('footer-extras')`, this would overwrite
anything that another template in the chain was passing via the section
of this same name. What I didn't understand was that even though an
individual page would've been extending a layout that extended another
layout, the page would've had access to the original layout's section
of the same name directly.

I only just today encountered a situation that produced this issue. In
light of this, I've decided to update the section names that, for
instance, the page layout uses inside of the master layout's similarly-
named sections. This will prevent accidentally overwriting anything that
the page layout needs to include on every page.

Same applies for the splash layout.