Have bootstrap.js initialize frontend
This is what this file was originally meant for, so it will handle this
for us from now on. The app.js file will only be an entry point from now
on; this is what it was meant for.

As part of this, the global Axios import and config that already existed
in bootstrap.js will be used in place of importing Axios individually
within the Vue components that make use of it.

Also as part of this, the global Vue import and initializing of the
global event bus will happen in the bootstrap.js file.

Lastly, some dependencies like underscore and popper are no longer
referenced in this file. I don't plan on using these, so they're not
worth importing.