Tweak base Page method
I added a method to Tests\Browser\Pages\Page for registering page components, but it didn't have a way to allow inheriting page classes to register additional components.

This requires that an array of component instances be passed to it for registration, including the defaults registered by the base page class.