Generate fake data directly instead of DTOs
It turned out that generating DTOs was going to put them into the shape that the application already wanted when the GithubService expected the data in the shape that GitHub itself is supplying. Generating DTOs would've forced the application to need to convert back to the shape of the data expected by the GithubService.

Given all of this, the new factory class will generate fake data arrays directly.